feat(tauri): typed commands + ask pipeline + progress Channel (P3)
Wire the Rust engine into Tauri 2 as typed (tauri-specta) commands:
- engine/classify.rs: the async `ask` pipeline (route -> trusted-retrieve ->
  extract -> solve -> gate -> record episode). FAIL-CLOSED: returns a plain
  Outcome (abstain on every error); never Err/panic to the UI. Store lock is
  taken only in synchronous closures and never held across an .await (the
  !Send rusqlite guard).
- lib.rs commands: ask(question, as_of_date, on_progress: Channel<String>),
  submit_correction, approve, list_topics, recent_episodes, list_strategy_items
  (+ greet kept). Store managed as State<Mutex<SqliteStore>> at app_data_dir.
- specta::Type derives on the engine types; EngineError as a typed command
  error; helpful/harmful counters i32 (specta forbids i64 export).
- src/lib/bindings.ts regenerated by the export_bindings cargo test (6 new
  typed commands + Channel).

check:all GREEN (biome + tsc + cargo fmt/clippy/check/test 42/42). Advisor-gated:
ask is fail-closed (classify_ask maps every error to abstain; Ok(_) is the only
return); lock never crosses await. App.tsx/capabilities unchanged (UI is P4).

+++ src-tauri/src/engine/learn.rs
@@ -35,8 +35,8 @@ pub struct Draft {
     pub tags: Vec<String>,
     pub trust_state: TrustState,
     pub provenance: Provenance,
-    pub helpful: i64,
-    pub harmful: i64,
+    pub helpful: i32,
+    pub harmful: i32,
 }
 
 impl Draft {
